+++ docs/CHANGELOG.md
@@ -4,7 +4,7 @@
 - Add `CfdiUtils\Certificado\SerialNumber::asDecimal()`
 - Fix `CfdiUtils\Cleaner\Cleaner` since internal `DOMDocument` can be null
 - Allow attributes `CfdiUtils\Elements\Cfdi33\Comprobante::getCfdiRelacionados`
-
+- Do not use `CfdiUtils\CadenaOrigen\DefaultLocations` at any place of the project
 
 # Version 2.4.4 2018-05-11
 - FIX: Unable to load a PEM file using filename on windows (Closes #33)